app/requirements.txt
aiogram==3.1.0
aiogram3-di==1.0.9
asyncpg==0.29.0
pydantic-settings==2.1.0
redis==5.0.1
sqlalchemy==2.0.23
